requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.1.4
django-bootstrap4==22.3
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.3.0
python-dotenv==0.21.0
pytz==2022.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
zipp==3.11.0
